:recycle: [Refactor] card 타입 정리 파일 위치 변경
