+++ src/page/driver/costHandler/mockData.ts
@@ -1,4 +1,4 @@
-import { ChipType } from '../../../components/card/type';
+import { ChipType } from '../../../types/cardTypes';
 
 export interface UserData {
   users: {
